requirements.txt
sqlalchemy[asyncio] 
asyncpg 
psycopg2-binary
pyjwt
flask
